Add Delete method

+++ AAD_SPRING/12_SpringDataJPA/src/main/java/lk/ijse/spring/service/impl/CustomerServiceImpl.java
@@ -26,7 +26,11 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public void deleteCustomer(String id) {
-        repo.deleteById(id);
+        if (repo.existsById(id)){
+            repo.deleteById(id);
+        }else{
+            throw new RuntimeException("Please check the Customer ID.. No Such Customer..!");
+        }
     }
 
     @Override
